+++ crates/macros/src/common/macros.rs
@@ -301,4 +301,8 @@ impl<'l> Macro<'l> {
 
         attrs
     }
+
+    pub fn is_untagged(&self) -> bool {
+        self.args.serde.untagged || self.serde_args.untagged
+    }
 }