+++ crates/macros/src/config/mod.rs
@@ -68,6 +68,7 @@ impl ToTokens for ConfigMacro<'_> {
         // Generate the partial implementation
         let partial_name = format_ident!("Partial{}", cfg.name);
         let partial_attrs = cfg.get_partial_attrs();
+        let is_untagged = cfg.is_untagged();
         let partial = cfg.type_of.generate_partial(
             &partial_name,
             &partial_attrs,
@@ -76,7 +77,7 @@ impl ToTokens for ConfigMacro<'_> {
         );
         let partial_default_impl = cfg
             .type_of
-            .generate_partial_default_impl(&partial_name, cfg.generics);
+            .generate_partial_default_impl(&partial_name, cfg.generics, is_untagged);
 
         tokens.extend(quote! {
             #partial