+++ crates/schematic/tests/generator_test.rs
@@ -57,6 +57,9 @@ struct GenConfig {
     /// **Nested** field.
     #[setting(nested)]
     nested: AnotherConfig,
+    /// Flattened field...
+    #[setting(flatten)]
+    flattened: HashMap<String, serde_json::Value>,
 
     // Types
     date: chrono::NaiveDate,
@@ -73,7 +76,7 @@ struct GenConfig {
     version_req: semver::VersionReq,
     json_value: serde_json::Value,
     toml_value: Option<toml::Value>,
-    yaml_value: serde_yml::Value,
+    yaml_value: serde_norway::Value,
     indexmap: IndexMap<String, String>,
     indexset: Option<IndexSet<String>>,
 }
@@ -283,6 +286,62 @@ mod template_json {
 
         assert_snapshot!(fs::read_to_string(file).unwrap());
     }
+
+    #[test]
+    fn only_fields() {
+        let sandbox = create_empty_sandbox();
+        let file = sandbox.path().join("schema.json");
+
+        create_template_generator()
+            .generate(
+                &file,
+                JsoncTemplateRenderer::new({
+                    let mut options = create_template_options();
+                    options.only_fields.push("float32".into());
+                    options.only_fields.push("string".into());
+                    options
+                }),
+            )
+            .unwrap();
+
+        assert_snapshot!(fs::read_to_string(file).unwrap());
+    }
+
+    #[test]
+    fn custom_values() {
+        let sandbox = create_empty_sandbox();
+        let file = sandbox.path().join("schema.json");
+
+        create_template_generator()
+            .generate(
+                &file,
+                JsoncTemplateRenderer::new({
+                    let mut options = create_template_options();
+                    options.custom_values.insert(
+                        "expandArrayPrimitive".into(),
+                        Schema::array(ArrayType::new(Schema::integer(IntegerType::new_unsigned(
+                            IntegerKind::Usize,
+                            123456,
+                        )))),
+                    );
+                    options.custom_values.insert(
+                        "nested.opt".into(),
+                        Schema::union(UnionType::new_any([
+                            Schema::string(StringType::new("custom nested value")),
+                            Schema::null(),
+                        ])),
+                    );
+                    options.custom_values.insert(
+                        "string".into(),
+                        Schema::string(StringType::new("custom value")),
+                    );
+                    options
+                }),
+            )
+            .unwrap();
+
+        assert_snapshot!(fs::read_to_string(file).unwrap());
+    }
 }
 
 #[cfg(all(feature = "renderer_template", feature = "pkl"))]
@@ -321,7 +380,7 @@ mod template_toml {
     }
 }
 
-#[cfg(all(feature = "renderer_template", feature = "yml"))]
+#[cfg(all(feature = "renderer_template", feature = "yaml"))]
 mod template_yaml {
     use super::*;
     use schematic::schema::*;